Add files via upload
Climate change is one of the biggest current global challenges, becoming increasingly relevant year upon year. There is a range of research which shows that people are, over time, becoming more concerned the effects of on the planet by this phenomenon. The purpose of this report is to see whether socio- demographic factors can impact a person’s attitudes towards climate change mitigation.